% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr.R
\name{geom_pie}
\alias{geom_pie}
\title{Draw a pie chart}
\usage{
geom_pie(
  mapping = NULL,
  data = NULL,
  stat = "pie_rescale",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Draw a pie chart
}
