% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr.R
\name{pie_slice}
\alias{pie_slice}
\title{Draw a slice}
\usage{
pie_slice(
  theta,
  theta_0 = 0,
  x_0 = 0.5,
  y_0 = 0.5,
  radius = 0.5,
  gp = NULL,
  n_points = 100
)
}
\arguments{
\item{n_points}{}
}
\value{

}
\description{
Draw a slice
}
